Allow formatting the demangled symbol name without the trailing hash.
This changeset allows for formatting the symbol name without the trailing
hash. The hash is useful for Rust's linking situation, but not particularly
useful to humans, and possibly harmful in cases where we want to use
function names for aggregation (like crash reporting).

Since the existing code is just using the `Display` trait for string conversion,
I made it support the "alternate" format modifier. I don't know if there's
a more sensible way to handle this. I could instead implement a separate
method on the type and have the `Display` implementation call that, if
you prefer.