Add an executable command line interface
This adds nothing to the library itself,
but makes it useable as a command-line utility.
Due to issues in cargo, you must add "--features binary"
to actualy compile the binary.
If you don't specify it, it either won't compile/install,
or it'll crash cargo.

By default the 'hashes' are excluded from the demangled names,
as that's probably what users want. However, it can be explicitly
enabled with the '-h' option.