Apply the recursion limit to the parser instead of just to the printer.
Raphael Salas found more stack overflows involving very large symbols that
failed while the parser was trying to skip over tokens. These are fixed by
moving the recursion limiter from the printer to the parser, and updating the
printer to use the parser's limiter.

The failing cases are added as unit tests.