Support searching for and demangling symbols
This adds an API for demangling a stream of data which consists of a mix
of mangled and non-mangled symbols. This is gated behind a new Cargo
feature ("std") since the new API uses I/O traits and allocates memory.

This API is broadly useful for any downstream users that are processing
streams with interleaved symbols (e.g., perf script output) and don't
want to filter down ahead of time. Previously the best option for this
(to my knowledge) was either shelling out to the `rustfilt` binary or
copying an implementation much like it.

The implementation added in this commit is roughly 2x faster than that
in the rustfilt binary today; if this is accepted moving rustfilt to use
this will make sense.